Unicode
[unicode::utf8::toxunicode(..)] Re-implemented and added overloading for 8/16/32 bit unsigned integers + unit tests

+++ src/unicode/unicode.cpp
@@ -42,7 +42,7 @@ std::ostream &blogator::unicode::normalize( std::ostream &os, const std::u32stri
 std::ostream &blogator::unicode::normalize( std::ostream &os, const std::string & u8str ) {
     for( auto c : u8str ) {
         if( unicode::ascii::iscntrl( c ) || c == '\'' || c == '\"' || c == '\\' || c == 0x00 ) {
-            os<< unicode::utf8::toxunicode( c );
+            os<< unicode::utf8::toxunicode( (uint8_t) c );
         } else {
             os << c;
         }
